Refactor Applications notifications as split files
